create dashboard module
